refactor(frontend): split API client into shared transport and features
- move Cookie fetch, CSRF handling, and ApiError normalization to shared/api
- split auth, tasks, sso, saml, and users API functions by feature
- move auth and task query hooks under feature modules
- delete legacy hooks and cookie-api-client compatibility entrypoints
- update UI imports to use feature query and API modules directly
- document the no compatibility wrapper policy for frontend refresh
