+++ frontend/src/components/tasks/task-create-dialog.tsx
@@ -15,7 +15,7 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
-import { useCookieCreateTodo } from '@/hooks/use-cookie-todo-queries';
+import { useCookieCreateTodo } from '@/features/tasks/queries';
 import { useUIStore } from '@/stores';
 import { Loader2 } from 'lucide-react';
