+++ frontend/src/components/tasks/task-edit-dialog.tsx
@@ -17,7 +17,7 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
 import { Checkbox } from '@/components/ui/checkbox';
-import { useCookieUpdateTodo } from '@/hooks/use-cookie-todo-queries';
+import { useCookieUpdateTodo } from '@/features/tasks/queries';
 import { useUIStore } from '@/stores';
 import { TodoResponse } from '@/types';
 import { Loader2 } from 'lucide-react';
