--- frontend/src/hooks/index.ts
@@ -1,4 +1,0 @@
-/** Main hooks export file (cookie-based only) */
-export * from './use-cookie-auth-queries';
-export * from './use-cookie-todo-queries';
-export * from './use-sso-login';
