+++ frontend/src/shared/api/index.ts
@@ -1,0 +1,6 @@
+export {
+  ApiError,
+  CookieApiClient,
+  cookieApiClient,
+  isApiError,
+} from './http-client';
